+++ src/PerfectBreakfast.API/Controllers/V1/DailyOrderController.cs
@@ -51,11 +51,5 @@ namespace PerfectBreakfast.API.Controllers.V1
             return response.IsError ? HandleErrorResponse(response.Errors) : Ok(response.Payload);
         }
 
-        [HttpPut]
-        public async Task<IActionResult> AutoUpdateDailyOrer(DateTime date)
-        {
-            var response = await _dailyOrderService.AutoUpdate(date);
-            return response.IsError ? HandleErrorResponse(response.Errors) : Ok(response.Payload);
-        }
     }
 }
