+++ src/PerfectBreakfast.Application/Interfaces/IDailyOrderService.cs
@@ -10,7 +10,6 @@ namespace PerfectBreakfast.Application.Interfaces
         public Task<OperationResult<List<DailyOrderResponse>>> GetDailyOrders();
         public Task<OperationResult<Pagination<DailyOrderResponse>>> GetDailyOrderPaginationAsync(int pageIndex = 0, int pageSize = 10);
         public Task<OperationResult<DailyOrderResponse>> UpdateDailyOrder(Guid id, UpdateDailyOrderRequest updateDailyOrderRequest);
-        public Task<OperationResult<DailyOrderResponse>> AutoUpdate(DateTime dateTime);
-        public Task<OperationResult<DailyOrderResponse>> AutoCreate(DateTime dateTime);
+
     }
 }
