+++ src/PerfectBreakfast.Application/Interfaces/IManagementService.cs
@@ -1,0 +1,7 @@
+﻿namespace PerfectBreakfast.Application.Interfaces;
+
+public interface IManagementService
+{
+    public Task AutoUpdateDailyOrderAfter4PM();
+    public Task AutoCreateDailyOrderEachDay1AM();
+}
