+++ src/PerfectBreakfast.Infrastructure/DenpendencyInjection.cs
@@ -6,6 +6,7 @@ using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
 using PerfectBreakfast.Application.Interfaces;
 using PerfectBreakfast.Application.Services;
+using PerfectBreakfast.Infrastructure.BackgroundJobServices;
 using PerfectBreakfast.Infrastructure.MailServices;
 using PerfectBreakfast.Infrastructure.Payments;
 using System.Reflection;
@@ -20,6 +21,8 @@ public static class DenpendencyInjection
         services.AddScoped<ICurrentTime, CurrentTime>();
         services.AddTransient<JWTService>();
         services.AddTransient<IMailService, MailService>();
+        services.AddScoped<IManagementService, ManagementService>();
+
         // ATTENTION: if you do migration please check file README.md
         services.AddDbContext<AppDbContext>(options =>
         {
